feat: Improve task description handling in 'Today' view
Introduces the ability to toggle the visibility of parent task descriptions within the 'Today' view. This change also refactors the description display logic in `TaskItem` to make it more reusable and lays the groundwork for potential inline editing of parent task descriptions in the future.

Additionally, a new `CalendarX2Icon` has been added to the icons library for future use.